+++ src/containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/RepeatPicker/RepeatPicker.tsx
@@ -1,21 +1,12 @@
 import { Dispatch, SetStateAction, useEffect, useState } from "react";
 import TopNavigationBar from "@components/layouts/common/TopNavigationBar";
-import RepeatRadioGroup from "@containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/RepeatPicker/radio/RepeatRadioGroup";
-import AllDay from "@containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/RepeatPicker/repeat/AllDay";
-import Week from "@containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/RepeatPicker/repeat/Week";
-import Month from "@containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/RepeatPicker/repeat/Month";
-import Year from "@containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/RepeatPicker/repeat/Year";
-import EndDate from "@containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/RepeatPicker/period/EndDate";
-import RadioLabel from "@containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/RepeatPicker/radio/RadioLabel";
-import RepetitionCount from "@containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/RepeatPicker/period/RepetitionCount";
+import RepeatRadioGroup from "@containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/RepeatPicker/components/radio/RepeatRadioGroup";
 import ThickDivider from "@components/common/ThickDivider";
 import RepeatInput from "@containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/RepeatInput";
 import { Box } from "@mui/material";
-import { UpdateStateInterface } from "@type/common";
 import { useScheduleForm } from "@containers/home/ScheduleDrawer/hooks/useScheduleForm";
-import { INIT_PERIOD, INIT_REPEAT } from "constants/schedule";
-import { SchedulePeriod, ScheduleRepeat } from "@type/schedule";
-import moment from "moment";
+import RepeatContainer from "./containers/RepeatContainer/RepeatContainer";
+import PeriodContainer from "./containers/PeriodContainer/PeriodContainer";
 
 export interface RepeatPickerProps {
   setIsRepeatPickerOpen: Dispatch<SetStateAction<boolean>>;
@@ -26,14 +17,8 @@ function RepeatPicker({ setIsRepeatPickerOpen }: RepeatPickerProps) {
   const [repeatType, setRepeatType] = useState<string>(
     scheduleForm ? scheduleForm.repeat.kind_type : "",
   );
-  const [repeat, setRepeat] = useState<ScheduleRepeat>(
-    INIT_REPEAT(moment(scheduleForm ? scheduleForm.start_date : "")),
-  );
   const [periodType, setPeriodType] = useState<string>(
     scheduleForm ? scheduleForm.repeat.kind_type : "",
-  );
-  const [period, setPeriod] = useState<SchedulePeriod>(
-    INIT_PERIOD(moment(scheduleForm ? scheduleForm.start_date : "")),
   );
 
   useEffect(() => {
@@ -44,20 +29,6 @@ function RepeatPicker({ setIsRepeatPickerOpen }: RepeatPickerProps) {
   useEffect(() => {
     updatePeriod({ target: { id: "period", value: periodType } });
   }, [periodType]);
-
-  const changeRepeat = (state: UpdateStateInterface) => {
-    updateRepeat(state);
-    setRepeat((pre) => {
-      return { ...pre, [state.target.id]: state.target.value };
-    });
-  };
-
-  const changePeriod = (state: UpdateStateInterface) => {
-    updatePeriod(state);
-    setPeriod((pre) => {
-      return { ...pre, [state.target.id]: state.target.value };
-    });
-  };
 
   return (
     <>
@@ -77,13 +48,7 @@ function RepeatPicker({ setIsRepeatPickerOpen }: RepeatPickerProps) {
             value={repeatType}
             handleChange={(v) => setRepeatType(v)}
           >
-            <AllDay repeatType={repeatType} />
-
-            <Week repeatType={repeatType} handleChangeOption={changeRepeat} />
-
-            <Month repeatType={repeatType} handleChangeOption={changeRepeat} />
-
-            <Year repeatType={repeatType} />
+            <RepeatContainer repeatType={repeatType} />
           </RepeatRadioGroup>
 
           <ThickDivider />
@@ -95,11 +60,7 @@ function RepeatPicker({ setIsRepeatPickerOpen }: RepeatPickerProps) {
             value={periodType}
             handleChange={(v: string) => setPeriodType(v)}
           >
-            <RadioLabel value="is_repeat_again" label="계속 반복" />
-
-            <RepetitionCount periodType={periodType} />
-
-            <EndDate handleChangeOption={changePeriod} />
+            <PeriodContainer periodType={periodType} />
           </RepeatRadioGroup>
         </>
       )}
