+++ src/containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/RepeatPicker/components/radio/RepeatRadioGroup.stories.tsx
@@ -8,7 +8,14 @@ const meta = {
   component: RepeatRadioGroup,
   tags: ["autodocs"],
   args: {
-    value: "is_repeat_again",
+    value: "day",
+    children: (
+      <>
+        <RadioLabel value="day" label="매일" />
+        <RadioLabel value="week" label="매주" />
+        <RadioLabel value="month" label="매달" />
+      </>
+    ),
   },
   argTypes: {},
 } satisfies Meta<typeof RepeatRadioGroup>;
@@ -18,15 +25,12 @@ export default meta;
 export const Default = (args: RepeatRadioGroupProps) => {
   return (
     <div style={{ width: "500px" }}>
-      <RepeatRadioGroup {...args}>
-        <RadioLabel value="is_repeat_again" label="계속 반복" />
-        <RadioLabel value="repeat_number_time" label={"일정 반복 횟수"} />
-      </RepeatRadioGroup>
+      <RepeatRadioGroup {...args} />
     </div>
   );
 };
 
-export const Dynamics = () => {
+export const Dynamic = () => {
   const [value, setValue] = useState("is_repeat_again");
   return (
     <div style={{ width: "500px" }}>
@@ -34,9 +38,11 @@ export const Dynamics = () => {
         value={value}
         handleChange={(value: string) => setValue(value)}
       >
-        <RadioLabel value="is_repeat_again" label="계속 반복" />
-        <RadioLabel value="repeat_number_time" label={"일정 반복 횟수"} />
-        <RadioLabel value="repeat_end_line" label={"종료 날짜"} />
+        <>
+          <RadioLabel value="is_repeat_again" label="계속 반복" />
+          <RadioLabel value="repeat_number_time" label={"일정 반복 횟수"} />
+          <RadioLabel value="repeat_end_line" label={"종료 날짜"} />
+        </>
       </RepeatRadioGroup>
     </div>
   );
