+++ src/containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/RepeatPicker/containers/PeriodContainer/index.tsx
@@ -1,0 +1,3 @@
+import PeriodContainer from "./PeriodContainer";
+
+export default PeriodContainer;
