+++ src/containers/home/ScheduleDrawer/pages/ScheduleFormPage/components/RepeatPicker/containers/RepeatContainer/repeat/RepeatInputLabel.tsx
@@ -1,4 +1,4 @@
-import InputLabel from "../radio/RadioLabel/labels/InputLabel";
+import InputLabel from "../../../components/radio/RadioLabel/labels/InputLabel";
 import { useScheduleForm } from "@containers/home/ScheduleDrawer/hooks/useScheduleForm";
 import { useEffect, useState } from "react";
